[wip] Added a test util, going to refactor server soon
