+++ propolis-server/Cargo.toml
@@ -29,3 +29,8 @@ structopt = { version = "0.3", default-features = false }
 propolis = { path = "../propolis" }
 propolis-client = { path = "../propolis-client" }
 uuid = "0.8"
+
+[dev-dependencies]
+hex = "0.4.3"
+reqwest = { version = "0.11", default-features = false, features = ["rustls-tls"] }
+ring = "0.16"
